Move safeProgressRatio to a shared helpers folder
