+++ src/components/HrTools/Shared/helpers/safeProgressRatio.test.ts
