+++ src/components/HrTools/Shared/helpers/safeProgressRatio.ts
